Cycle.cs: renamed setCyclePosition() to setStartPosition() which makes more sense changed type of speed from float to int to save some memory
GameSettings.cs:
changed type of PLAYER_QUANTITY from int to byte to save some memory

Player.cs:
changed type of player id from int to byte to save some memory

Round.cs:
changed type of mapMirror from float[,] to byte[,]
moved playerlist From FuFiCycles.cs to Round.cs
implemented a check if all cycles are crashed to find out when to start a new round

+++ Core/GameSettings.cs
@@ -6,7 +6,7 @@ namespace Fusee.FuFiCycles.Core {
 		public static FuFiCycles INSTANCE;
 		public static List<Round> ROUNDS = new List<Round>();
 		public static bool SHOW_MINIMAP = true;
-		public static int PLAYER_QUANTITY = 2;
+		public static byte PLAYER_QUANTITY = 2;
 		public static int MAP_SIZE;
 		public static float4x4 SCENE_SCALE = float4x4.CreateScale(0.04f);
 	}
